Move clock choosing to tlog_rec
Move clock choosing to tlog_rec to allow passing it to rate-limiting
writer in the following commits.
